+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 * Change confirm delete rows actions on async.
 
+### Added
+* Available mime types check in `flexberry-file` component.
+
 ## [2.4.0] - 2020-05-18
 ### Added
 * Attributes with selectors for testing in the generated application.